Fix a bug in trampoline prepare.
The offset of the variable the MEM_WR instruction would touch was begin
computed unproperly in the case a relocation was applying to the instruction.
In fact, no offset would be found since the linker would compute it afterwards,
however the trampoline structure was not being updated.

To fix it a new relocation entry is added in case the current instrumented
instruction has a referenced symbol to it.

Signed-off-by: tankwalker <cingodvd@gmail.com>