Fix RIP-relative addressing
Trampoline sequence did not take into account RIP-relative addressing
during the computation of the final variable's address to look for,
causing to addressing the wrong memeory region.

A new flag has been added to trampoline.h in order to handle this case,
along with the relative check on the trampolie64.S sequence.

Signed-off-by: tankwalker <cingodvd@gmail.com>