Fix instruction's flags bug in insn_entry creation
The code that prepares the trampoline stack was failing to initialize
the proper value of flags field of the insturction's structure.
In particular the generic internal x86 flags values was being inherited,
instead of the proper trampoline's ones, defined in 'trampoline.h'.

A new code block has been added in 'reverse-x86.c' to correctly compute
instruction's flags which specifically regard the addressing mode used.

Signed-off-by: tankwalker <cingodvd@gmail.com>