Adding new flag in XML config file.
In the XML configuration file the instruction's tag has only a flag
to indicate which type of instruction to instrument; however could be
necessary to designate a set of flags to skip.

A new tag's property has been added to the supported schema, which allows
to specify instruction's flags to skip.

Signed-off-by: tankwalker <cingodvd@gmail.com>