EFLAG status instruction added in reverse generation
The EFLAG register must be handle by hijacker since it makes room for the
reverse trampoline on the stack. Doing so EFLAG register will be touched
possibly creating a misbehavior in case a compare instruction follows the
instrumented-trampoline instrucion.

Hijacker will add a new couple of instructions in order to preserver ELFAG
value across the trampoline execution.

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>
Cherry-pick from master