Fixing an alignment issue in trampoline
System ABI v requires the stack to be aligned to double-word, which
was not ensured by the present assembly code in trampoline64.S
The fix must be verified against future alteration of the stack by
hijacker during the instrumentation phase.

Actual fix is to add a couple of 'sub' and 'add' instructions with a
hardwired value to compensate the 'pushfw' and 'popfw' 2-bytes operations.

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>