smtracer replacement for vptracker
The new 'smtracer' preset is an improved version of 'vptracker'
which gives the user better control on the overall accuracy
vs overhead trade-off of instrumentation. It still accepts
an instrumentation score threshold for basic blocks, as well
as a chunk size---a chunk can be anything from a byte cell to
a huge page---and a flag to ignore stack accesses. Additionally,
it can be configured to instrument a strict fraction of memory
accesses within a basic block. This portion is computed on the
basis of a user-defined accuracy factor.

Signed-off-by: Simone Economo <simone.economo@gmail.com>