Fixed access equivalence and distance computation issues
This commit fixes two issues, one related to address equivalence
and one related to the computation of access distances. Relocation
types and addends are now taken into account in both cases, since
they contribute to the final displacement value injected into an
expression. Moreover, a bug has been fixed which considered two
expressions with the same template but a missing quantity to be
potentially considered as equivalent.

Signed-off-by: Simone Economo <economo@dis.uniroma1.it>