Fixed incorrect section size expansion in the emit stage
This commit fixes a bug occurring when ELF sections are
written to the output file. When the span of the requested
write exceeds the total allocated size, the latter gets
doubled to accomodate this new write request.
However, it may be the case that the condition still holds.
In that case, a segmentation fault will arise because
there is still insufficient space for the write.
Therefore, we need to iteratively evluate the condition
until it becomes true. In practice, this translates into
a replacement of an IF statement with a WHILE statement.

Signed-off-by: Simone Economo <economo@dis.uniroma1.it>