TLS buffer structure change for the smtracer preset
This commit introduces a change in the structure of the TLS buffer
populate by the smtracer preset. Specifically, each entry in the
buffer is now 24 bytes large. It is composed by three 8-byte fields:
the chunk address, the check access count and the basic block id.
By doing this, this commit also fixes a bug which led to add the
basic block id to an existing entry multiple times.

Signed-off-by: Simone Economo <economo@dis.uniroma1.it>