Refactoring Hijacker's utility function and data structures
This is the first of a (long) sequence of commits to refactor
Hijacker's current codebase. It targets both utility functions
and utility data structures.

As for functions, they have been polished and properly documented
using DocGen syntax. A few header files have been merged, too.

As for data structures, this commit introduces (and documents)
two old companions: balanced binary search trees and graphs.
The difference between this version and previous one is that
while the latter was tightly coupled with the IBR, the former
can be used whenever appropriate in the code. For example,
graphs can now be used to implement both CFGs and the CFG, as
well as to fuel future data-flow analysis tasks. A similar
reasoning holds for binary trees, which could be used to index
heterogeneous IBR structures.

Note that this commit breaks compatibility with old code and
actually prevent it from compiling. This will be fixed graudally
as soon as the new codebase starts to stabilize again.

Signed-off-by: Simone Economo <economo@dis.uniroma1.it>