Refactoring Hijacker's Intermediate Binary Representation
This commit brings a tentative refactoring of Hijacker's IBR,
which closely reflects the approach taken in the current
stable version.

The most relevant change compared to the historical IBR is the
decoupling of symbols and actual relocations, to more easily
manage references from sections to symbols.

Note that this refactoring is by no means complete and will be
likely subject to drastic changes in the immediate future.
Nevertheless, it hopefully makes for a good starting point to
improve things.

Signed-off-by: Simone Economo <economo@dis.uniroma1.it>