Implemented insertion and removal of instructions from the IBR
This commit finalizes the APIs to insert and remove instructions
from the Intermediate Binary Representation. Specifically, it allows
to derive new instructions either from a mnemonic string or a raw
sequence of bytes. Multiple instructions can be inserted at once
with a single call, before or after a pivotal instruction.

The only known issue left concerns blocks and how to deal with
pivot instructions that act as delimiters. This is left for future
discussion, once basic blocks are rewieved at the IBR level.

Signed-off-by: Simone Economo <economo@dis.uniroma1.it>