Implemented removal/replacement of multiple instructions
Currently 'instr_replace' returns a pointer to the first inserted
instruction, but it also populates (if desired) the value of a passed
pointer with the address of the last inserted instruction.
This API is not stable, as it doesn't reflect what is returned by other
chain-manipulating functions. Therefore, it requires further reasoning.

Signed-off-by: Simone Economo <economo@dis.uniroma1.it>