Loading and writing object files
This commit lays the basis for loading and writing object files. Currently,
it requires libbfd as the underlying object file handling library, but it
is flexbile enough so as to allow other (better) parsers to be used in the future.

Ohter improvements concern APIs for inserting/removing IBR objects, as well as
commonly used data structures (mainly, linked lists).

Signed-off-by: Simone Economo <economo@dis.uniroma1.it>