This commit introduces substantial changes to the way `smtracer` works.
- The flushpoint logic has been rewritten and simplified and now operates
at the basic block level, to achieve correct tracing results. Specifically,
the user-defined routine is invoked by the end of the last traced memory-
accessing instruction within the same basic block. This way, the correctness
of tracing is still preserved across function calls (recall that when a
CALL is encountered, Hijacker splits the current block into two new blocks).
Note that this causes much higher overheads than before, so a better (future)
solution would be to restore the previous flushing logic but with dynamic
indexes in place of statically-assigned ones (which were wrong).

- A new selection algorithm has been devised in order to correctly treat
zero-distance pairs of expressions. It is based on a freeze-unfreeze
mechanism which only considers one expression among all those at distance
zero: namely, the one with the highest absolute value. As soon as this
expression is found, the other expressions are frozen and are no more
considered for selection. If there's still space left in the Knapsack and
there are no other expressions to pick, the algorithm unfreezes the frozen
ones and begins a new iteration. This guarantees both vertical accuracy
(i.e., it always picks expressions with the highest access count) and
horizontal accuracy (i.e., it is best effort in terms of how many portions
of the address space are traced).

- The preset now expects an overhead parameter rather than an accuracy
one. The reason behind this change is that the user is typically more
interested in putting a bound on the overhead than on the accuracy.