- `smtracer` now collects stats for the current experiment. It accepts two new parameters: a name for the test and a boolean which indicates that indeed it must gather statistical measures. The former is internally used to generate unique names for the statistical reports produced by Hijacker by the end of the instrumentation.
- To approximate the expected outcome of instrumenting an arbitrary
percent of instructions within a basic block, a naive error-propagation
algorithm has been implemented. It scans blocks linearly in the order
they materialize in the code. For each block the number of memory-referencing
instructions to instrument is computed using both `floor` and `ceiling`,
along with absolute errors. The specific function that Hijacker will use is
the one that minimizes the sum of the previously-accumulated absolute
error and the one introduced by the function. After a choice has been made,
the new accumulated error gets propagated to the next block.

- Minor bugfixes and changes