Fix that patches trampoline symbol duplication
The symbol to the `trampoline` is recreated each time with a different
name the `trampoline_prepare` function is called; however this behavior
will affect the linking stage, where the linker will not found any of
the duplicated symbols.

This fix will prevent to create a-priori a new symbols each time the
trampolie must be prepared by searching if it is yet present.

This issue comes from the actual behavior of the `symbol_create`
API is to create a new symbol, independently of the fact that this
symbol actualy exists, since could be a WEAK one, so they MUST cohexist
event though they have the same name (but different binding).

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>