Fix for the write of '.rodata' section
There is a typo in the emitter code such that the
`.rodata` section will be populated with spurious
data, since it is used the old section's payload
data pointer instead of the `content` variable,
which holds the newly formatted data to write.

The spurious data will be merged in the linking
phase with the relocation information and generates
poisoned references (e.g. for the jump tables).

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>