Fix issue in .rodata relocation shift update
In the `shift_instruction_addresses` function, the relocation
in `.rodata` are not updated because of typo that forces to
use the wrong relocation symbol list, starting from the instruction.

In particular the `reference` varible is used instead of the
`pointedby` one. The first holds the relocation that apply
to `.text`, wheread the `pointedby` holds the list of the
relocations that apply in data sections.

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>